+++ src/simulation/vizard/_GeneralModuleFiles/vizStructures.h
@@ -171,7 +171,7 @@ GenericSensor
     std::vector<int> color;             //!< [] (optional) RGBA as values between 0 and 255, multiple colors can be populated in this field and will be assigned to the additional mode (Modes 0 and 1 will use the 0th color, Mode 2 will use the color indexed to 1, etc.  If 2 colors are provided, then the vector should have size 8 (2x4 color channels)
     std::string label = "";             //!< [] (optional) string to display on sensor label
     ReadFunctor<DeviceCmdMsgPayload> genericSensorCmdInMsg;   //!< [-] (Optional)  incoming sensor cmd state msg
-    int genericSensorCmd = 0;           //!< [int] (optional) sensor cmd value, if cmd input msg is connected, then this is set from the message
+    uint64_t genericSensorCmd = 0;      //!< [int] (optional) sensor cmd value, if cmd input msg is connected, then this is set from the message
 
 }GenericSensor;
